check that ctx exits first

+++ gita/__main__.py
@@ -132,7 +132,7 @@ def f_group(args: argparse.Namespace):
         ctx = utils.get_context()
         for name in args.to_ungroup:
             del groups[name]
-            if str(ctx.stem) == name:
+            if ctx and str(ctx.stem) == name:
                 ctx.unlink()
         utils.write_to_groups_file(groups, 'w')
     elif cmd == 'add':
